Add Audio vorbis conversion Support (#6)
* Add Audio File extraction support

This adds Audio support for NiceTube that will get the best audio version, then extract it and convert to oggvorbis for use in resonite as a soundclip.

* add oggvorbis to webpanel

Update the web panel to support the new audio only oggvorbis option.

* Update counter for audio

Separates the counter so i can see if the audio conversion is helpful for people.

* bump to 1.1

Happy with its state to remove the beta label